// almora-shop-ui/src/gql/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: { input: any; output: any };
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { input: any; output: any };
  /** The `Money` scalar type represents monetary values and supports signed double-precision fractional values as specified by [IEEE 754](https://en.wikipedia.org/wiki/IEEE_floating_point). */
  Money: { input: number; output: number };
  /** The `Upload` scalar type represents a file upload. */
  Upload: { input: any; output: any };
};
// ...
export type Asset = Node & {
  __typename?: "Asset";
  createdAt: Scalars["DateTime"]["output"];
  customFields?: Maybe<Scalars["JSON"]["output"]>;
  fileSize: Scalars["Int"]["output"];
  focalPoint?: Maybe<Coordinate>;
  height: Scalars["Int"]["output"];
  id: Scalars["ID"]["output"];
  mimeType: Scalars["String"]["output"];
  name: Scalars["String"]["output"];
  preview: Scalars["String"]["output"];
  source: Scalars["String"]["output"];
  tags: Array<Tag>;
  type: AssetType;
  updatedAt: Scalars["DateTime"]["output"];
  width: Scalars["Int"]["output"];
};
// ...
export type Customer = Node & {
  __typename?: "Customer";
  addresses?: Maybe<Array<Address>>;
  createdAt: Scalars["DateTime"]["output"];
  customFields?: Maybe<Scalars["JSON"]["output"]>;
  emailAddress: Scalars["String"]["output"];
  firstName: Scalars["String"]["output"];
  id: Scalars["ID"]["output"];
  lastName: Scalars["String"]["output"];
  orders: OrderList;
  phoneNumber?: Maybe<Scalars["String"]["output"]>;
  title?: Maybe<Scalars["String"]["output"]>;
  updatedAt: Scalars["DateTime"]["output"];
  user?: Maybe<User>;
};
// ...
export type Discount = {
  __typename?: "Discount";
  adjustmentSource: Scalars["String"]["output"];
  amount: Scalars["Money"]["output"];
  amountWithTax: Scalars["Money"]["output"];
  description: Scalars["String"]["output"];
  type: AdjustmentType;
};
// ...
export type Fulfillment = Node & {
  __typename?: "Fulfillment";
  createdAt: Scalars["DateTime"]["output"];
  customFields?: Maybe<Scalars["JSON"]["output"]>;
  id: Scalars["ID"]["output"];
  lines: Array<FulfillmentLine>;
  method: Scalars["String"]["output"];
  state: Scalars["String"]["output"];
  /** @deprecated Use the `lines` field instead */
  summary: Array<FulfillmentLine>;
  trackingCode?: Maybe<Scalars["String"]["output"]>;
  updatedAt: Scalars["DateTime"]["output"];
};
// ...
export type HistoryEntryList = PaginatedList & {
  __typename?: "HistoryEntryList";
  items: Array<HistoryEntry>;
  totalItems: Scalars["Int"]["output"];
};
// ...
export type Node = {
  id: Scalars["ID"]["output"];
};
// ...
export type Order = Node & {
  __typename?: "Order";
  /** An order is active as long as the payment process has not been completed */
  active: Scalars["Boolean"]["output"];
  billingAddress?: Maybe<OrderAddress>;
  /** A unique code for the Order */
  code: Scalars["String"]["output"];
  /** An array of all coupon codes applied to the Order */
  couponCodes: Array<Scalars["String"]["output"]>;
  createdAt: Scalars["DateTime"]["output"];
  currencyCode: CurrencyCode;
  customFields?: Maybe<Scalars["JSON"]["output"]>;
  customer?: Maybe<Customer>;
  discounts: Array<Discount>;
  fulfillments?: Maybe<Array<Fulfillment>>;
  history: HistoryEntryList;
  id: Scalars["ID"]["output"];
  lines: Array<OrderLine>;
  /**
   * The date & time that the Order was placed, i.e. the Customer
   * completed the checkout and the Order is no longer "active"
   */
  orderPlacedAt?: Maybe<Scalars["DateTime"]["output"]>;
  payments?: Maybe<Array<Payment>>;
  /** Promotions applied to the order. Only gets populated after the payment process has completed. */
  promotions: Array<Promotion>;
  shipping: Scalars["Money"]["output"];
  shippingAddress?: Maybe<OrderAddress>;
  shippingLines: Array<ShippingLine>;
  shippingWithTax: Scalars["Money"]["output"];
  state: Scalars["String"]["output"];
  /**
   * The subTotal is the total of all OrderLines in the Order. This figure also includes any Order-level
   * discounts which have been prorated (proportionally distributed) amongst the items of each OrderLine.
   * To get a total of all OrderLines which does not account for prorated discounts, use the
   * sum of `OrderLine.discountedLinePrice` values.
   */
  subTotal: Scalars["Money"]["output"];
  /** Same as subTotal, but inclusive of tax */
  subTotalWithTax: Scalars["Money"]["output"];
  /**
   * Surcharges are arbitrary modifications to the Order total which are neither
   * ProductVariants nor discounts resulting from applied Promotions. For example,
   * one-off discounts based on customer interaction, or surcharges based on payment
   * methods.
   */
  surcharges: Array<Surcharge>;
  /** A summary of the taxes being applied to this Order */
  taxSummary: Array<OrderTaxSummary>;
  /** Equal to subTotal plus shipping */
  total: Scalars["Money"]["output"];
  totalQuantity: Scalars["Int"]["output"];
  /** The final payable amount. Equal to subTotalWithTax plus shippingWithTax */
  totalWithTax: Scalars["Money"]["output"];
  type: OrderType;
  updatedAt: Scalars["DateTime"]["output"];
};

export type OrderHistoryArgs = {
  options?: InputMaybe<HistoryEntryListOptions>;
};

export type OrderAddress = {
  __typename?: "OrderAddress";
  city?: Maybe<Scalars["String"]["output"]>;
  company?: Maybe<Scalars["String"]["output"]>;
  country?: Maybe<Scalars["String"]["output"]>;
  countryCode?: Maybe<Scalars["String"]["output"]>;
  customFields?: Maybe<Scalars["JSON"]["output"]>;
  fullName?: Maybe<Scalars["String"]["output"]>;
  phoneNumber?: Maybe<Scalars["String"]["output"]>;
  postalCode?: Maybe<Scalars["String"]["output"]>;
  province?: Maybe<Scalars["String"]["output"]>;
  streetLine1?: Maybe<Scalars["String"]["output"]>;
  streetLine2?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type OrderLine = Node & {
  __typename?: "OrderLine";
  createdAt: Scalars["DateTime"]["output"];
  customFields?: Maybe<Scalars["JSON"]["output"]>;
  /** The price of the line including discounts, excluding tax */
  discountedLinePrice: Scalars["Money"]["output"];
  /** The price of the line including discounts and tax */
  discountedLinePriceWithTax: Scalars["Money"]["output"];
  /**
   * The price of a single unit including discounts, excluding tax.
   *
   * If Order-level discounts have been applied, this will not be the
   * actual taxable unit price (see `proratedUnitPrice`), but is generally the
   * correct price to display to customers to avoid confusion
   * about the internal handling of distributed Order-level discounts.
   */
  discountedUnitPrice: Scalars["Money"]["output"];
  /** The price of a single unit including discounts and tax */
  discountedUnitPriceWithTax: Scalars["Money"]["output"];
  discounts: Array<Discount>;
  featuredAsset?: Maybe<Asset>;
  fulfillmentLines?: Maybe<Array<FulfillmentLine>>;
  id: Scalars["ID"]["output"];
  /** The total price of the line excluding tax and discounts. */
  linePrice: Scalars["Money"]["output"];
  /** The total price of the line including tax but excluding discounts. */
  linePriceWithTax: Scalars["Money"]["output"];
  /** The total tax on this line */
  lineTax: Scalars["Money"]["output"];
  order: Order;
  /** The quantity at the time the Order was placed */
  orderPlacedQuantity: Scalars["Int"]["output"];
  productVariant: ProductVariant;
  /**
   * The actual line price, taking into account both item discounts _and_ prorated (proportionally-distributed)
   * Order-level discounts. This value is the true economic value of the OrderLine, and is used in tax
   * and refund calculations.
   */
  proratedLinePrice: Scalars["Money"]["output"];
  /** The proratedLinePrice including tax */
  proratedLinePriceWithTax: Scalars["Money"]["output"];
  /**
   * The actual unit price, taking into account both item discounts _and_ prorated (proportionally-distributed)
   * Order-level discounts. This value is the true economic value of the OrderItem, and is used in tax
   * and refund calculations.
   */
  proratedUnitPrice: Scalars["Money"]["output"];
  /** The proratedUnitPrice including tax */
  proratedUnitPriceWithTax: Scalars["Money"]["output"];
  /** The quantity of items purchased */
  quantity: Scalars["Int"]["output"];
  taxLines: Array<TaxLine>;
  taxRate: Scalars["Float"]["output"];
  /** The price of a single unit, excluding tax and discounts */
  unitPrice: Scalars["Money"]["output"];
  /** Non-zero if the unitPrice has changed since it was initially added to Order */
  unitPriceChangeSinceAdded: Scalars["Money"]["output"];
  /** The price of a single unit, including tax but excluding discounts */
  unitPriceWithTax: Scalars["Money"]["output"];
  /** Non-zero if the unitPriceWithTax has changed since it was initially added to Order */
  unitPriceWithTaxChangeSinceAdded: Scalars["Money"]["output"];
  updatedAt: Scalars["DateTime"]["output"];
};
// ...
export type Product = Node & {
  __typename?: "Product";
  assets: Array<Asset>;
  collections: Array<Collection>;
  createdAt: Scalars["DateTime"]["output"];
  customFields?: Maybe<Scalars["JSON"]["output"]>;
  description: Scalars["String"]["output"];
  enabled: Scalars["Boolean"]["output"];
  facetValues: Array<FacetValue>;
  featuredAsset?: Maybe<Asset>;
  id: Scalars["ID"]["output"];
  languageCode: LanguageCode;
  name: Scalars["String"]["output"];
  optionGroups: Array<ProductOptionGroup>;
  slug: Scalars["String"]["output"];
  translations: Array<ProductTranslation>;
  updatedAt: Scalars["DateTime"]["output"];
  /** Returns a paginated, sortable, filterable list of ProductVariants */
  variantList: ProductVariantList;
  /** Returns all ProductVariants */
  variants: Array<ProductVariant>;
};
// ...
export type ProductVariant = Node & {
  __typename?: "ProductVariant";
  assets: Array<Asset>;
  createdAt: Scalars["DateTime"]["output"];
  currencyCode: CurrencyCode;
  customFields?: Maybe<Scalars["JSON"]["output"]>;
  facetValues: Array<FacetValue>;
  featuredAsset?: Maybe<Asset>;
  id: Scalars["ID"]["output"];
  languageCode: LanguageCode;
  name: Scalars["String"]["output"];
  options: Array<ProductOption>;
  price: Scalars["Money"]["output"];
  priceWithTax: Scalars["Money"]["output"];
  product: Product;
  productId: Scalars["ID"]["output"];
  sku: Scalars["String"]["output"];
  stockLevel: Scalars["String"]["output"];
  taxCategory: TaxCategory;
  taxRateApplied: TaxRate;
  translations: Array<ProductVariantTranslation>;
  updatedAt: Scalars["DateTime"]["output"];
};
// ...
export type ActiveOrderFieldsFragment = {
  __typename?: "Order";
  id: string;
  code: string;
  state: string;
  active: boolean;
  totalQuantity: number;
  totalWithTax: number;
  currencyCode: CurrencyCode;
  shippingWithTax: number;
  subTotalWithTax: number;
  couponCodes: Array<string>;
  lines: Array<{
    __typename?: "OrderLine";
    id: string;
    quantity: number;
    linePriceWithTax: number;
    discountedLinePriceWithTax: number;
    unitPriceWithTax: number;
    productVariant: {
      __typename?: "ProductVariant";
      id: string;
      name: string;
      sku: string;
      priceWithTax: number;
      currencyCode: CurrencyCode;
      productId: string;
      product: {
        __typename?: "Product";
        slug: string;
        featuredAsset?: {
          __typename?: "Asset";
          id: string;
          preview: string;
        } | null;
      };
      featuredAsset?: {
        __typename?: "Asset";
        id: string;
        preview: string;
      } | null;
    };
  }>;
} & { " $fragmentName"?: "ActiveOrderFieldsFragment" };